backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
psycopg2-binary
python-dotenv
